backend/requirements.txt
django
djangorestframework
django-cors-headers
django-allauth
psycopg2-binary
djangorestframework-simplejwt
pandas
spacy
nltk
sentence-transformers
scikit-learn
PyPDF2
tqdm
selenium
beautifulsoup4 
lxml
reportlab

#python -m spacy download fr_core_news_sm
